advanced_python_oop_udemy_course/app_7_automating_emails/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
cssutils==1.0.2
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
lxml==4.6.2
numpy==1.26.4
openpyxl==3.0.5
pandas==1.5.3
premailer==3.7.0
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.15.0
urllib3==1.26.2
yagmail==0.14.245